Minor tweaks on the TODO list.

+++ src/pages/todo/TodoListItem.js
@@ -1,7 +1,9 @@
+import React from "react";
 import { makeStyles, Button, Checkbox } from "@material-ui/core";
 import { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { todoActions } from "../../store/todo-list";
+import TodoEntryEditForm from "./forms/TodoEntryEditForm";
 
 const useStyles = makeStyles({
   todoListItem: {
@@ -20,7 +22,7 @@ const useStyles = makeStyles({
   done: {
     order: 2,
   },
-  "h2.done": {
+  doneTitle: {
     textDecoration: "line-through",
   },
 });
@@ -30,16 +32,29 @@ const TodoListItem = (props) => {
 
   const { id, title, content, completed } = props.item;
 
+  const { showConfirmation } = props;
+  const { draggableProps, dragHandleProps, innerRef } = props.provided;
+
   const [check, setCheck] = useState(completed);
+  const [editing, setEditing] = useState(false);
   const dispatch = useDispatch();
 
-  const editEntryHandler = () => {};
-
-  const removeEntryHandler = () => {
-    dispatch(todoActions.removeItem(id));
+  const editEntryHandler = () => {
+    setEditing(true);
   };
 
-  const toggleHandler = (event) => {
+  const closeEditEntryHandler = () => {
+    setEditing(false);
+  };
+
+  const removeEntryHandler = () => {
+    const confirmation = showConfirmation
+      ? window.confirm("Are you sure you want to delete this entry?")
+      : true;
+    confirmation && dispatch(todoActions.removeItem(id));
+  };
+
+  const toggleCompletionHandler = (event) => {
     const checkbox = event.target;
     if (!checkbox) {
       return;
@@ -51,22 +66,41 @@ const TodoListItem = (props) => {
     dispatch(action);
   };
 
-  return (
-    <li className={`${classes.todoListItem} ${check ? classes.done : ""}`}>
+  const todoEntryContent = editing ? (
+    <TodoEntryEditForm item={props.item} onClose={closeEditEntryHandler} />
+  ) : (
+    <>
       <div className={classes.todoListItem__content}>
-        <Checkbox onChange={toggleHandler} checked={check} />
-        <h2 className={check ? classes.done : ""}>{title}</h2>
+        <Checkbox onChange={toggleCompletionHandler} checked={check} />
+        <h2 className={check ? classes.doneTitle : ""}>{title}</h2>
       </div>
       <div>
         <p>{content}</p>
       </div>
+    </>
+  );
+
+  return (
+    <li
+      className={`${classes.todoListItem} ${check ? classes.done : ""}`}
+      {...draggableProps}
+      {...dragHandleProps}
+      ref={innerRef}
+    >
+      {todoEntryContent}
       <div className={classes.todoListItem__controls}>
-        <Button color="primary" onClick={editEntryHandler}>
-          Edit entry
-        </Button>
-        <Button color="secondary" onClick={removeEntryHandler}>
-          Remove entry
-        </Button>
+        {editing ? (
+          ""
+        ) : (
+          <>
+            <Button color="primary" onClick={editEntryHandler}>
+              Edit entry
+            </Button>
+            <Button color="secondary" onClick={removeEntryHandler}>
+              Remove entry
+            </Button>
+          </>
+        )}
       </div>
     </li>
   );
